refactor identity service
