+++ app/services/HttpIdentityService.scala
@@ -1,22 +1,24 @@
 package services
 
+import java.net.URI
+
 import cats.data.EitherT
-import com.gu.identity.play.{IdMinimalUser, IdUser, PrivateFields, PublicFields}
-import play.api.libs.ws.{BodyWritable, InMemoryBody, WSClient, WSRequest, WSResponse}
+import cats.implicits._
+import com.google.common.net.InetAddresses
+import com.gu.identity.play.{IdMinimalUser, IdUser}
+import config.Identity
+import models.identity.UserIdWithGuestAccountToken
+import models.identity.requests.CreateGuestAccountRequestBody
+import models.identity.responses.{GuestRegistrationResponse, UserResponse}
+import monitoring.SafeLogger
+import monitoring.SafeLogger._
+import play.api.libs.json.Json
+import play.api.libs.ws.{WSClient, WSRequest, WSResponse}
 import play.api.mvc.RequestHeader
 
 import scala.concurrent.duration._
-import cats.implicits._
-import java.net.URI
-
-import akka.util.ByteString
-import com.google.common.net.InetAddresses
-import config.Identity
-import monitoring.SafeLogger
-import monitoring.SafeLogger._
-import play.api.libs.json.{Json, Reads, Writes}
-import scala.util.Try
 import scala.concurrent.{ExecutionContext, Future}
+import scala.util.Try
 
 object IdentityServiceEnrichers {
 
@@ -50,92 +52,6 @@ object IdentityService {
         apiClientToken = config.apiClientToken
       )
     }
-  }
-}
-
-// Models a valid request to /guest
-//
-// Example request:
-// =================
-// {
-//   "email": "a@b.com",
-//   "publicFields": {
-//     "displayName": "a"
-//   }
-// }
-case class CreateGuestAccountRequestBody(primaryEmailAddress: String, publicFields: PublicFields)
-object CreateGuestAccountRequestBody {
-
-  def guestDisplayName(email: String): String = email.split("@").headOption.getOrElse("Guest User")
-
-  def fromEmail(email: String): CreateGuestAccountRequestBody = CreateGuestAccountRequestBody(email, PublicFields(Some(guestDisplayName(email))))
-
-  implicit val writesCreateGuestAccountRequestBody: Writes[CreateGuestAccountRequestBody] = Json.writes[CreateGuestAccountRequestBody]
-  implicit val bodyWriteable: BodyWritable[CreateGuestAccountRequestBody] = BodyWritable[CreateGuestAccountRequestBody](
-    transform = body => InMemoryBody(ByteString.fromString(Json.toJson(body).toString)),
-    contentType = "application/json"
-  )
-
-}
-
-// Models the response of successfully creating a guest account.
-//
-// Example response:
-// =================
-// {
-//   "status": "ok",
-//   "guestRegistrationRequest": {
-//     "token": "83e41c1d-458d-49c0-b469-ddc263507034",
-//     "userId": "100000190",
-//     "timeIssued": "2018-02-28T14:46:01Z"
-//   }
-// }
-case class GuestRegistrationResponse(
-    guestRegistrationRequest: GuestRegistrationResponse.GuestRegistrationRequest
-)
-
-// A userId with an optional guest account registration toke (created when a guest account is registered)
-case class UserIdWithGuestAccountToken(userId: String, guestAccountRegistrationToken: Option[String])
-
-object UserIdWithGuestAccountToken {
-  def fromGuestRegistrationResponse(guestRegistrationResponse: GuestRegistrationResponse): UserIdWithGuestAccountToken =
-    UserIdWithGuestAccountToken(
-      guestRegistrationResponse.guestRegistrationRequest.userId,
-      guestRegistrationResponse.guestRegistrationRequest.token
-    )
-}
-
-object GuestRegistrationResponse {
-  implicit val readsGuestRegistrationResponse: Reads[GuestRegistrationResponse] = Json.reads[GuestRegistrationResponse]
-  case class GuestRegistrationRequest(token: Option[String], userId: String)
-
-  object GuestRegistrationRequest {
-    implicit val readsGuestRegistrationRequest: Reads[GuestRegistrationRequest] = Json.reads[GuestRegistrationRequest]
-  }
-}
-
-// Models the response of successfully looking up user details via email address.
-//
-// Example response:
-// =================
-// {
-//   "status": "ok",
-//   "user": {
-//     "id": "100000190",
-//     "dates": {
-//       "accountCreatedDate": "2018-02-28T14:46:01Z"
-//     }
-//   }
-// }
-case class UserResponse(user: UserResponse.User)
-
-object UserResponse {
-
-  implicit val readsUserResponse: Reads[UserResponse] = Json.reads[UserResponse]
-  case class User(id: String)
-
-  object User {
-    implicit val readsUser: Reads[User] = Json.reads[User]
   }
 }
