+++ app/services/StubIdentityService.scala
@@ -3,6 +3,7 @@ package services
 import cats.data.EitherT
 import cats.implicits._
 import com.gu.identity.play.{IdMinimalUser, IdUser, PrivateFields, PublicFields}
+import models.identity.UserIdWithGuestAccountToken
 import monitoring.SafeLogger
 import play.api.mvc.RequestHeader
